Bug #48461: Downgrade spamasassin in preup, improved version check

+++ base/univention-updater/script/preup.sh
@@ -649,14 +649,16 @@ fi
 # which is incompatible with spamassassin + dependencies from UCS 4.3-0
 check_and_downgrade_spamassassin_version ()
 {
-	spamassassin_dpkg_ver="$(dpkg-query -W -f '${Version}\n' "spamassassin")"
-	if LC_ALL=C dpkg --compare-versions "$spamassassin_dpkg_ver" gt "3.4.0-6"; then
-		echo spamassassin version "$spamassassin_dpkg_ver" >>"$UPDATER_LOG"
-		if is_ucr_true update43/override_spamassassin_downgrade; then
-			echo "WARNING: update43/override_spamassassin_downgrade is set to true. Skipped as requested."
-		else
-			apt-get install spamassassin=3.4.0-6 >>"$UPDATER_LOG" 2>&1
-			ucr set repository/online/component/4.2-5-errata=false >>"$UPDATER_LOG" 2>&1
+	if dpkg -l 'spamassassin' 2>/dev/null | grep -q ^ii; then
+		spamassassin_dpkg_ver="$(dpkg-query -W -f '${Version}\n' "spamassassin")"
+		if LC_ALL=C dpkg --compare-versions "$spamassassin_dpkg_ver" gt "3.4.0-6"; then
+			echo spamassassin version "$spamassassin_dpkg_ver" >>"$UPDATER_LOG"
+			if is_ucr_true update43/override_spamassassin_downgrade; then
+				echo "WARNING: update43/override_spamassassin_downgrade is set to true. Skipped as requested."
+			else
+				apt-get install spamassassin=3.4.0-6 >>"$UPDATER_LOG" 2>&1
+				ucr set repository/online/component/4.2-5-errata=false >>"$UPDATER_LOG" 2>&1
+			fi
 		fi
 	fi
 }
